Move IdSugarRing function
