+++ src/Exec_PrepareForLeap.cpp
@@ -138,322 +138,10 @@ int Exec_PrepareForLeap::LoadPdbResNames(std::string const& fnameIn)
 
 
 
-/// Recursive function for following bonds of an atom to a target atom
-static void FollowBonds(int atm, Topology const& topIn, int idx, std::vector<int>& ring_atoms, int tgt_atom, std::vector<bool>& Visited, bool& found)
-{
-  Visited[atm] = true;
-  int rnum = topIn[atm].ResNum();
-  //for (int i = 0; i != idx; i++) // DEBUG
-  //  mprintf("\t"); // DBEUG
-  //mprintf("At atom %s\n", topIn.ResNameNumAtomNameNum(atm).c_str()); // DEBUG
-  ring_atoms[idx] = atm;
-  // Assume we have started at the target atom
-  if (idx > 0 && atm == tgt_atom) {
-    found = true;
-    return;
-  }
-  // Follow all atoms bonded to this atom
-  for (Atom::bond_iterator bat = topIn[atm].bondbegin(); bat != topIn[atm].bondend(); ++bat)
-  {
-    if (topIn[*bat].ResNum() == rnum &&
-        topIn[*bat].Element() == Atom::CARBON &&
-        !Visited[*bat])
-    {
-      FollowBonds( *bat, topIn, idx+1, ring_atoms, tgt_atom, Visited, found );
-      if (found) return;
-    }
-  }
-}
 
 // -----------------------------------------------
 // -----------------------------------------------
-/** \return true if residue is on the list of residues with missing heteroatoms. */
-static bool res_missing_het(int rnum, Topology const& topIn) {
-  Residue const& res = topIn.Res(rnum);
-  //mprintf("DEBUG: TGT %s %i '%c' '%c'\n", *(res.Name()), res.OriginalResNum(), res.Icode(), res.ChainId());
-  for (std::vector<Residue>::const_iterator het = topIn.MissingHet().begin();
-                                            het != topIn.MissingHet().end(); ++het)
-  {
-    //mprintf("DEBUG: HET %s %i '%c' '%c'\n", *(het->Name()), het->OriginalResNum(), het->Icode(), het->ChainId());
-    if ( het->OriginalResNum() == res.OriginalResNum() &&
-         het->Icode()          == res.Icode() &&
-         het->ChainId()        == res.ChainId() &&
-         het->Name()           == res.Name() )
-      return true;
-  }
-  return false;
-}
 
-/** Identify sugar oxygen, anomeric and ref carbons, and ring atoms. */
-Exec_PrepareForLeap::Sugar Exec_PrepareForLeap::IdSugarRing(int rnum, Topology const& topIn)
-const
-{
-  Residue const& res = topIn.Res(rnum);
-  bool residue_missing_atoms = res_missing_het(rnum, topIn);
-  if (residue_missing_atoms)
-    mprintf("\tResidue '%s' is missing atoms.\n", topIn.TruncResNameOnumId(rnum).c_str());
-
-  // Determine candidates for ring oxygen atoms. 
-  Iarray potentialRingStartAtoms;
-  // This array will hold the 2 flanking carbons for potential ring oxygen atoms.
-  std::vector<std::pair<int,int>> ringAtomCarbons;
-  for (int at = res.FirstAtom(); at != res.LastAtom(); at++)
-  {
-    Atom const& currentAtom = topIn[at];
-    // Try to identify the sugar ring oxygen. Candidate atoms are oxygens
-    // bonded to two carbon atoms in the same residue.
-    if (currentAtom.Element() == Atom::OXYGEN) {
-      if (currentAtom.Nbonds() > 1) {
-        Iarray c_atoms;
-        for (Atom::bond_iterator bat = currentAtom.bondbegin();
-                                 bat != currentAtom.bondend(); ++bat)
-        {
-          if (topIn[*bat].Element() == Atom::CARBON &&
-              topIn[*bat].ResNum() == rnum)
-            c_atoms.push_back( *bat );
-        }
-        if ( c_atoms.size() == 2 ) {
-          potentialRingStartAtoms.push_back( at );
-          ringAtomCarbons.push_back( std::pair<int,int>(c_atoms[0], c_atoms[1]) );
-        }
-      }
-    }
-  }
-
-  if (potentialRingStartAtoms.empty()) {
-    mprintf("Warning: Ring oxygen could not be identified for %s\n",
-            topIn.TruncResNameOnumId(rnum).c_str());
-    return Sugar(Sugar::MISSING_O, res.FirstAtom());
-  }
-
-  // Use the previously-set up AtomMap to help determine stereocenters
-  std::vector<bool> atomIsChiral;
-  atomIsChiral.reserve( res.NumAtoms() );
-  // Since we cannot be certain there will be hydrogens, cannot rely
-  // on the AtomMap chiral designations (which assumes hydrogens).
-  // Make a chiral center a carbon with at least 3 bonds, all must be
-  // to different kinds of atoms.
-  int resat = res.FirstAtom();
-  for (int iat = 0; iat != res.NumAtoms(); iat++, resat++)
-  {
-    bool chiral = false;
-    if (topIn[resat].Element() == Atom::CARBON && topIn[resat].Nbonds() > 2) {
-      if (debug_ > 0)
-        mprintf("DEBUG: Atom '%s' potential chiral\n", topIn.TruncResAtomNameNum(resat).c_str());
-      chiral = true;
-      for (Atom::bond_iterator bat1 = topIn[resat].bondbegin();
-                               bat1 != topIn[resat].bondend(); ++bat1)
-      {
-        std::string unique1 = myMap_[*bat1].Unique();
-        for (Atom::bond_iterator bat2 = bat1 + 1; bat2 != topIn[resat].bondend(); ++bat2)
-        {
-          std::string unique2 = myMap_[*bat2].Unique();
-          if (unique1 == unique2) {
-            // At least two of the atoms bonded to this atom look the same. Not chiral.
-            if (debug_ > 1)
-              mprintf("DEBUG: unique strings match %s='%s' %s='%s'\n", *(topIn[*bat1].Name()), unique1.c_str(), *(topIn[*bat2].Name()), unique2.c_str());
-            chiral = false;
-            break;
-          }
-        } // END inner loop over bonded atoms
-        if (!chiral) break;
-      } // END outer loop over bonded atoms
-    } // END atom is carbon with > 2 bonds
-    
-    atomIsChiral.push_back( chiral );
-    if (debug_ > 0)
-      mprintf("DEBUG: Atom '%s' isChiral= %i\n",
-              topIn.TruncResAtomNameNum(resat).c_str(),
-              (int)atomIsChiral.back());
-  }
-
-  // Using the potential ring start atoms, see if we can actually complete
-  // a ring. Identify important atoms as well.
-  // This will indicate ring direction
-//  int ring_direction = 0;
-  // Ring end atom is the last atom in the ring
-  int ring_end_atom = -1;
-  // The anomeric carbon is the carbon that was part of the carbonyl group
-  // in the straight chain. It is therefore typically the carbon with fewer
-  // bonds to other carbons.
-  int anomeric_atom = -1;    // e.g. C1
-  // The anomeric reference carbon is the stereocenter farthest from the
-  // anomeric carbon in the ring.
-  int ano_ref_atom = -1;     // e.g. C5
-  // Ring oxygen atom
-  int ring_oxygen_atom = -1; // e.g. O5
-  // This will hold the index of the highest stereocenter, e.g. C5
-  int highest_stereocenter = -1;
-  // This will hold ring atoms, not including the ring oxygen.
-  std::vector<Iarray> Ring_Atoms;
-  // This will hold carbon chain atoms starting from the anomeric carbon
-  Iarray carbon_chain;
-
-  // Out of the potential ring start atoms, see which ones are actually
-  // part of a ring. Potential ring start atoms only have 2 bonds,
-  // each one to a carbon.
-  int ring_atom_idx = -1;
-  std::vector<std::pair<int,int>>::iterator catoms = ringAtomCarbons.begin();
-  for (Iarray::const_iterator ringat = potentialRingStartAtoms.begin();
-                              ringat != potentialRingStartAtoms.end();
-                            ++ringat, ++catoms)
-  {
-    if (debug_ > 0)
-      mprintf("DEBUG: Ring start '%s'\n", topIn.ResNameNumAtomNameNum(*ringat).c_str());
-    // Mark all atoms as visited except this residue (minus the ring start).
-    std::vector<bool> Visited( topIn.Natom(), true );
-    for (int at = res.FirstAtom(); at != res.LastAtom(); at++)
-      if (at != *ringat)
-        Visited[at] = false;
-    Iarray ring_atoms( topIn.Res(rnum).NumAtoms(), -1 );
-    // Will be set true if complete ring can be found
-    bool ring_complete = false;
-
-    // Since we have already established that *ringat is an oxygen bonded
-    // to two carbons, just start at the first carbon to see if we can
-    // get to the second carbon.
-    int c_beg, c_end;
-    if (catoms->first < catoms->second) {
-      c_beg = catoms->first;
-      c_end = catoms->second;
-    } else {
-      c_beg = catoms->second;
-      c_end = catoms->first;
-    }
-    // Try to ascertain which carbon might be the anomeric carbon (i.e. the
-    // carbon that originally started the chain). Tie goes to lower index.
-    int c_beg_bonds_to_C = 0;
-    for (Atom::bond_iterator bat = topIn[c_beg].bondbegin(); bat != topIn[c_beg].bondend(); ++bat)
-      if (topIn[*bat].Element() == Atom::CARBON)
-        c_beg_bonds_to_C++;
-    int c_end_bonds_to_C = 0;
-    for (Atom::bond_iterator bat = topIn[c_end].bondbegin(); bat != topIn[c_end].bondend(); ++bat)
-      if (topIn[*bat].Element() == Atom::CARBON)
-        c_end_bonds_to_C++;
-    if (debug_ > 0)
-      mprintf("DEBUG:\t(%s bonds to C= %i, %s bonds to C = %i)\n", // DEBUG
-              topIn.ResNameNumAtomNameNum(c_beg).c_str(), c_beg_bonds_to_C,
-              topIn.ResNameNumAtomNameNum(c_end).c_str(), c_end_bonds_to_C);
-    if (c_beg_bonds_to_C <= c_end_bonds_to_C) {
-      anomeric_atom = c_beg;
-      ring_end_atom = c_end;
-//      ring_direction = 1;
-    } else {
-      anomeric_atom = c_end;
-      ring_end_atom = c_beg;
-//      ring_direction = -1;
-    }
-//    mprintf("DEBUG: Potential Ring direction= %i\n", ring_direction);
-    catoms->first = anomeric_atom;
-    catoms->second = ring_end_atom;
-
-    FollowBonds(anomeric_atom, topIn, 0, ring_atoms,
-                ring_end_atom, Visited, ring_complete);
-    if (debug_ > 0)
-      mprintf("DEBUG: Potential ring start atom %s, Ring complete = %i",
-              topIn.ResNameNumAtomNameNum(*ringat).c_str(), (int)ring_complete);
-    // Create empty array for ring atoms
-    Ring_Atoms.push_back(Iarray());
-    if (ring_complete) {
-      // Able to complete the cycle.
-      if (ring_atom_idx == -1)
-        ring_atom_idx = (int)(ringat - potentialRingStartAtoms.begin());
-      else {
-        mprinterr("Error: Multiple potential ring atoms: %s and %s\n",
-                  topIn.ResNameNumAtomNameNum(potentialRingStartAtoms[ring_atom_idx]).c_str(),
-                  topIn.ResNameNumAtomNameNum(*ringat).c_str());
-        return Sugar(Sugar::MULTIPLE_O, res.FirstAtom());
-      }
-      // Place the ring atoms into an array without the terminating -1
-      Iarray& RA = Ring_Atoms.back();
-      if (debug_ > 0) mprintf(" :"); // DEBUG
-      for (Iarray::const_iterator it = ring_atoms.begin(); it != ring_atoms.end(); ++it)
-      {
-        if (debug_ > 0) mprintf(" %i", *it + 1);
-        if (*it == -1) break;
-        RA.push_back( *it );
-      }
-    }
-    if (debug_ > 0) mprintf("\n"); // DEBUG
-  } // END loop over potential ring atoms
-  if (ring_atom_idx == -1) {
-    mprinterr("Error: Sugar ring oxygen could not be identified.\n");
-    return Sugar(Sugar::MISSING_O, res.FirstAtom());
-  }
-
-  ring_oxygen_atom = potentialRingStartAtoms[ring_atom_idx];
-  anomeric_atom = ringAtomCarbons[ring_atom_idx].first;
-  ring_end_atom = ringAtomCarbons[ring_atom_idx].second;
-  Iarray const& RA = Ring_Atoms[ring_atom_idx];
-
-  // Find anomeric reference atom. Start at ring end and work down to anomeric atom
-  for (Iarray::const_iterator arat = RA.end() - 1; arat != RA.begin(); --arat)
-    if (atomIsChiral[*arat - topIn.Res(rnum).FirstAtom()]) {
-      ano_ref_atom = *arat;
-      break;
-    }
-
-  // Get complete chain starting from the anomeric carbon
-  carbon_chain = RA;
-  if (FindRemainingChainCarbons(carbon_chain, ring_end_atom, topIn, rnum, RA)) {
-    mprinterr("Error: Could not find remaining chain carbons.\n");
-    return Sugar(Sugar::MISSING_CHAIN, res.FirstAtom());
-  }
-  if (debug_ > 0) {
-    mprintf("DEBUG: Complete carbon chain (from anomeric carbon):\n");
-    for (Iarray::const_iterator it = carbon_chain.begin(); it != carbon_chain.end(); ++it)
-      mprintf("\t\t%s\n", topIn.ResNameNumAtomNameNum(*it).c_str());
-  }
-  // See if there is chain prior to anomeric carbon
-  Iarray previous_chain;
-  if (FindRemainingChainCarbons(previous_chain, anomeric_atom, topIn, rnum, RA)) {
-    mprinterr("Error: Could not find previous chain carbons.\n");
-    return Sugar(Sugar::MISSING_CHAIN, res.FirstAtom());
-  }
-  if (!previous_chain.empty()) {
-    if (debug_ > 0) {
-      mprintf("DEBUG: Previous carbon chain (from anomeric carbon):\n");
-      for (Iarray::const_iterator it = previous_chain.begin(); it != previous_chain.end(); ++it)
-        mprintf("\t\t%s\n", topIn.ResNameNumAtomNameNum(*it).c_str());
-    }
-    for (Iarray::const_iterator it = carbon_chain.begin(); it != carbon_chain.end(); ++it)
-      previous_chain.push_back( *it );
-    carbon_chain = previous_chain;
-    if (debug_ > 0) {
-      mprintf("DEBUG: Complete carbon chain:\n");
-      for (Iarray::const_iterator it = carbon_chain.begin(); it != carbon_chain.end(); ++it)
-        mprintf("\t\t%s\n", topIn.ResNameNumAtomNameNum(*it).c_str());
-    }
-  } 
-  // Get the index of the highest stereocenter
-  for (Iarray::const_iterator it = carbon_chain.begin(); it != carbon_chain.end(); ++it)
-  {
-    if (atomIsChiral[*it - topIn.Res(rnum).FirstAtom()])
-      highest_stereocenter = *it;
-  }
-  if (debug_ > 0)
-    mprintf("DEBUG: Index of highest stereocenter: %s\n",
-            topIn.ResNameNumAtomNameNum(highest_stereocenter).c_str());
-
-  if (ano_ref_atom == -1) {
-    mprinterr("Error: Anomeric reference atom could not be identified.\n");
-    return Sugar(Sugar::MISSING_ANO_REF, ring_oxygen_atom, anomeric_atom, RA, carbon_chain);
-  }
-  if (highest_stereocenter == -1) {
-    mprinterr("Error: Highest stereocenter atom could not be identified.\n");
-    return Sugar(Sugar::MISSING_CONFIG, ring_oxygen_atom, anomeric_atom, RA, carbon_chain);
-  }
-
-//  if (!ring_complete || RA.empty() || ring_oxygen_atom == -1) {
-//    mprinterr("Error: Sugar ring atoms could not be identified.\n");
-//    stat = ID_ERR;
-//    return Sugar(res.FirstAtom());
-//  }
-  if (debug_ > 0)
-    mprintf("\t  Ring oxygen         : %s\n", topIn.ResNameNumAtomNameNum(ring_oxygen_atom).c_str());
-  return Sugar(ring_oxygen_atom, anomeric_atom, ano_ref_atom, highest_stereocenter,
-               RA, carbon_chain, residue_missing_atoms);
-}
 
 /** Change PDB atom names in residue to glycam ones. */
 int Exec_PrepareForLeap::ChangePdbAtomNamesToGlycam(std::string const& resCode, Residue const& res,
